Move AGP versions for perf tests to one class
