+++ subprojects/performance/src/performanceTest/groovy/org/gradle/performance/regression/corefeature/FileSystemWatchingPerformanceTest.groovy
@@ -22,7 +22,7 @@ import org.gradle.internal.os.OperatingSystem
 import org.gradle.performance.AbstractCrossVersionPerformanceTest
 import org.gradle.performance.annotations.RunFor
 import org.gradle.performance.annotations.Scenario
-import org.gradle.performance.fixture.IncrementalAndroidTestProject
+import org.gradle.performance.fixture.AndroidTestProject
 import org.gradle.performance.fixture.IncrementalTestProject
 import org.gradle.performance.fixture.TestProjects
 import org.gradle.test.fixtures.file.LeaksFileHandles
@@ -39,7 +39,6 @@ import static org.gradle.performance.results.OperatingSystem.WINDOWS
 )
 @LeaksFileHandles("The TAPI keeps handles to the distribution it starts open in the test JVM")
 class FileSystemWatchingPerformanceTest extends AbstractCrossVersionPerformanceTest {
-    private static final String AGP_TARGET_VERSION = "4.2"
     private static final String KOTLIN_TARGET_VERSION = new KotlinGradlePluginVersions().latests.last()
 
     def setup() {
@@ -88,8 +87,8 @@ class FileSystemWatchingPerformanceTest extends AbstractCrossVersionPerformanceT
 
     IncrementalTestProject findTestProjectAndSetupRunnerForFsWatching(boolean enableConfigurationCaching) {
         IncrementalTestProject testProject = TestProjects.projectFor(runner.testProject)
-        if (testProject instanceof IncrementalAndroidTestProject) {
-            IncrementalAndroidTestProject.configureForLatestAgpVersionOfMinor(runner, AGP_TARGET_VERSION)
+        if (testProject instanceof AndroidTestProject) {
+            AndroidTestProject.useLatestAgpVersion(runner)
             runner.args.add("-D${StartParameterBuildOptions.ConfigurationCacheProblemsOption.PROPERTY_NAME}=warn")
             runner.args.add("-DkotlinVersion=${KOTLIN_TARGET_VERSION}")
         }
